DRY compare plant flowers for sorting
