+++ controllers/sectionControllers.js
@@ -3,6 +3,7 @@
 /*******************************************/
 const express = require('express');
 const Garden = require('../models/garden');
+const ControllerHelper = require('../utils/controllerHelper');
 
 /*******************************************/
 /*****          Create Router          *****/
@@ -181,7 +182,11 @@ router.get('/sections/:id', (req, res) => {
             // Get section
             const section = foundGarden.sections.id(sectionId);
             // Render edit page
-            res.render('sections/show', { section, garden: foundGarden, username, loggedIn, userId });
+            res.render('sections/show', { 
+                section, 
+                garden: foundGarden, 
+                comparePlantedFlowers: ControllerHelper.comparePlantedFlowers,
+                username, loggedIn, userId });
         })
         .catch(err => {
             // Handle any errors
