+++ views/partials/header.ejs
@@ -6,7 +6,6 @@
     <title>My Flower Garden</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet" href="/styles.css">
-    <script src="/script.js"></script>
 </head>
 <body>
     <%- include('./nav') %>
